修正：在Sqlserver数据库中，结课考试成绩核算存在异常的问题

+++ Sourcecode/Song.ServiceImpls/ExaminationCom.cs
@@ -890,7 +890,7 @@ namespace Song.ServiceImpls
             }
             #endregion
             outXml = resXml.InnerXml;
-            return (double)score;
+            return score;
         }
         #endregion
         /// <summary>
